requirements.txt
playwright==1.44.0
google-genai>=1.0.0
typer==0.12.3
pydantic==2.7.1
rich==13.7.1
pyyaml>=6.0
aiofiles==23.2.1
aiosqlite==0.20.0
pillow>=10.0.0
python-dotenv>=1.0.0
